+++ NINA/Utility/Enum/BacklashCompensationModel.cs
@@ -21,23 +21,17 @@
 
 #endregion "copyright"
 
-using System.Threading;
-using System.Threading.Tasks;
+using System.ComponentModel;
 
-namespace NINA.Model.MyFocuser {
+namespace NINA.Utility.Enum {
 
-    public interface IFocuser : IDevice {
-        bool IsMoving { get; }
-        int MaxIncrement { get; }
-        int MaxStep { get; }
-        int Position { get; }
-        double StepSize { get; }
-        bool TempCompAvailable { get; }
-        bool TempComp { get; set; }
-        double Temperature { get; }
+    [TypeConverter(typeof(EnumDescriptionTypeConverter))]
+    public enum BacklashCompensationModel {
 
-        Task Move(int position, CancellationToken ct);
+        [Description("LblAbsolute")]
+        ABSOLUTE,
 
-        void Halt();
+        [Description("LblOvershoot")]
+        OVERSHOOT
     }
 }